+++ src/index.ts
@@ -1,13 +1,14 @@
 
 
 
-//  I M P O R T S
+///  I M P O R T
 
 import { parseFile as envFile } from "env-smart";
-import print from "@webb/console";
 import rootFile from "app-root-path";
 
-//  U T I L S
+///  U T I L
+
+import { deepmerge, messageDeveloper } from "./helper";
 
 interface LooseObject {
   [key: string]: any
@@ -18,15 +19,17 @@ const separatorRegex = /-|\.|·/g;
 
 
 
-//  E X P O R T
+///  E X P O R T
 
-export default (suppliedEnv?: string) => {
+export default vne;
+
+export function vne(suppliedEnv?: string) {
   const defaultObject: LooseObject = {};
-  const nestedObject: LooseObject = {};
+  let nestedObject: LooseObject = {};
   let _env: LooseObject = {};
 
   if (suppliedEnv) {
-    /// You can supply an absolute path to an environment file.
+    /// You can supply an absolute path to your environment file.
     _env = envFile(suppliedEnv);
   } else {
     /// By default, the file named ".env" relative to your project will be used.
@@ -42,30 +45,28 @@ export default (suppliedEnv?: string) => {
   Object.keys(_env).map((variable: LooseObject|string) => {
     /// Numbers are removed from variable name.
     const cleanVariable = variable.replace(/\d+/g, "");
+    const stringifiedVariable = String(variable);
 
     switch(true) {
-      case (
-        variable.match(numberedStringRegex) &&
-        variable.match(numberedStringRegex).length > 0
-      ):
+      case (variable.match(numberedStringRegex) && variable.match(numberedStringRegex).length > 0):
         if (!defaultObject[cleanVariable]) {
           /// If array does not exist, it will get created.
           defaultObject[cleanVariable] = [];
         }
 
-        defaultObject[cleanVariable].push(_env[String(variable)]);
+        defaultObject[cleanVariable].push(_env[stringifiedVariable]);
         break;
 
-      case (
-        variable.match(separatorRegex) &&
-        variable.match(separatorRegex).length > 0
-      ):
-        nestedObject[variable.split(separatorRegex)[0]] = nestedObject[variable.split(separatorRegex)[0]] || {};
-        nestedObject[variable.split(separatorRegex)[0]][variable.split(separatorRegex)[1]] = _env[String(variable)];
+      case (variable.match(separatorRegex) && variable.match(separatorRegex).length > 0):
+        /// Recursive object handling made simple.
+        nestedObject = deepmerge(
+          nestedObject,
+          variable.split(separatorRegex).reduceRight((value: any, key: string) => ({ [key]: value }), _env[stringifiedVariable])
+        );
         break;
 
       default:
-        defaultObject[String(variable)] = _env[String(variable)];
+        defaultObject[stringifiedVariable] = _env[stringifiedVariable];
         break;
     }
   });
@@ -75,11 +76,5 @@ export default (suppliedEnv?: string) => {
 
 
 
-//  H E L P E R
-
-function messageDeveloper(suppliedMessage: string) {
-  console.log(
-    print.magentaLine(print.black(" [vne] ")) +
-    print.invert(` ${suppliedMessage} `) + "\n"
-  );
-}
+/// DEBUG
+// console.log(vne("path/to/local/.env/to/debug/this/module"));